🚚 chore : UIImage 함수 위치 변경
