+++ Mukku/MukkuWidgets/HomeWidgets/WidgetsView/PhantomOfOperaItem.swift
@@ -1,17 +1,6 @@
 import SwiftUI
 import WidgetKit
 
-extension UIImage {
-  func resized(toWidth width: CGFloat, isOpaque: Bool = true) -> UIImage? {
-    let canvas = CGSize(width: width, height: CGFloat(ceil(width/size.width * size.height)))
-    let format = imageRendererFormat
-    format.opaque = isOpaque
-    return UIGraphicsImageRenderer(size: canvas, format: format).image {
-      _ in draw(in: CGRect(origin: .zero, size: canvas))
-    }
-  }
-}
-
 struct PhantomOfOperaItem: View {
     
     @Environment(\.widgetFamily) var widgetFamily
