+++ Mukku/MukkuWidgets/HomeWidgets/WidgetsView/UIImage.swift
@@ -1,0 +1,19 @@
+//
+//  UIImage.swift
+//  Mukku
+//
+//  Created by Diane on 2023/05/16.
+//
+
+import Foundation
+import SwiftUI
+extension UIImage {
+  func resized(toWidth width: CGFloat, isOpaque: Bool = true) -> UIImage? {
+    let canvas = CGSize(width: width, height: CGFloat(ceil(width/size.width * size.height)))
+    let format = imageRendererFormat
+    format.opaque = isOpaque
+    return UIGraphicsImageRenderer(size: canvas, format: format).image {
+      _ in draw(in: CGRect(origin: .zero, size: canvas))
+    }
+  }
+}
